+++ test/cask/observable/TestObservableQueue.cpp
@@ -13,37 +13,116 @@ using cask::Scheduler;
 using cask::scheduler::BenchScheduler;
 
 TEST(ObservableQueue, Empty) {
-    auto sched = Scheduler::global();
-    auto result = Observable<int>::empty()
+    auto sched = std::make_shared<BenchScheduler>();
+    auto fiber = Observable<int>::empty()
         ->queue(1)
         ->last()
-        .run(sched)
-        ->await();
+        .run(sched);
+    
+    sched->run_ready_tasks();
+    auto result = fiber->await();
 
     EXPECT_FALSE(result.has_value());
 }
 
 TEST(ObservableQueue, Value) {
-    auto sched = Scheduler::global();
-    auto result = Observable<int>::pure(123)
+    auto sched = std::make_shared<BenchScheduler>();
+    auto fiber = Observable<int>::pure(123)
         ->queue(1)
         ->last()
-        .run(sched)
-        ->await();
+        .run(sched);
+    
+    sched->run_ready_tasks();
+    auto result = fiber->await();
 
     ASSERT_TRUE(result.has_value());
     EXPECT_EQ(*result, 123);
 }
 
+TEST(ObservableQueue, ValueThenError) {
+    auto sched = std::make_shared<BenchScheduler>();
+    auto fiber = Observable<int,std::string>::pure(123)
+        ->concat(Observable<int,std::string>::raiseError("broke"))
+        ->queue(1)
+        ->last()
+        .failed()
+        .run(sched);
+    
+    sched->run_ready_tasks();
+    auto result = fiber->await();
+    EXPECT_EQ(result, "broke");
+}
+
+TEST(ObservableQueue, Never) {
+    auto sched = std::make_shared<BenchScheduler>();
+    auto fiber = Observable<int>::never()
+        ->queue(1)
+        ->last()
+        .run(sched);
+    
+    sched->run_ready_tasks();
+
+    EXPECT_EQ(sched->num_task_ready(), 0);
+    EXPECT_EQ(sched->num_timers(), 0);
+    EXPECT_FALSE(fiber->getValue().has_value());
+    EXPECT_FALSE(fiber->getError().has_value());
+    EXPECT_FALSE(fiber->isCanceled());
+
+    fiber->cancel();
+    sched->run_ready_tasks();
+    EXPECT_TRUE(fiber->isCanceled());
+}
+
+TEST(ObservableQueue, NeverEarlyCancel) {
+    auto sched = std::make_shared<BenchScheduler>();
+    auto fiber = Observable<int>::never()
+        ->queue(1)
+        ->last()
+        .run(sched);
+
+    EXPECT_EQ(sched->num_task_ready(), 1);
+    EXPECT_EQ(sched->num_timers(), 0);
+    EXPECT_FALSE(fiber->getValue().has_value());
+    EXPECT_FALSE(fiber->getError().has_value());
+    EXPECT_FALSE(fiber->isCanceled());
+
+    fiber->cancel();
+    sched->run_ready_tasks();
+    EXPECT_TRUE(fiber->isCanceled());
+}
+
+TEST(ObservableQueue, ValueThenNever) {
+    auto sched = std::make_shared<BenchScheduler>();
+    auto fiber = Observable<int,std::string>::pure(123)
+        ->concat(Observable<int,std::string>::never())
+        ->queue(1)
+        ->last()
+        .run(sched);
+    
+    sched->run_ready_tasks();
+
+    EXPECT_EQ(sched->num_task_ready(), 0);
+    EXPECT_EQ(sched->num_timers(), 0);
+    EXPECT_FALSE(fiber->getValue().has_value());
+    EXPECT_FALSE(fiber->getError().has_value());
+    EXPECT_FALSE(fiber->isCanceled());
+
+    fiber->cancel();
+    sched->run_ready_tasks();
+    EXPECT_TRUE(fiber->isCanceled());
+}
+
 TEST(ObservableQueue, ValuesLargerThanQueue) {
-    auto sched = Scheduler::global();
-    auto result = Observable<int>::fromVector({
+    auto sched = std::make_shared<BenchScheduler>();
+    auto fiber = Observable<int>::fromVector({
             0, 1, 2, 3, 4, 5
         })
         ->queue(1)
         ->take(10)
-        .run(sched)
-        ->await();
+        .run(sched);
+    
+    sched->run_ready_tasks();
+    auto result = fiber->await();
 
     ASSERT_EQ(result.size(), 6);
     EXPECT_EQ(result[0], 0);
@@ -55,13 +134,15 @@ TEST(ObservableQueue, ValuesLargerThanQueue) {
 }
 
 TEST(ObservableQueue, Error) {
-    auto sched = Scheduler::global();
-    auto result = Observable<int, std::string>::raiseError("broke")
+    auto sched = std::make_shared<BenchScheduler>();
+    auto fiber = Observable<int, std::string>::raiseError("broke")
         ->queue(1)
         ->last()
         .failed()
-        .run(sched)
-        ->await();
+        .run(sched);
+
+    sched->run_ready_tasks();
+    auto result = fiber->await();
 
     EXPECT_EQ(result, "broke");
 }
@@ -106,4 +187,52 @@ TEST(ObservableQueue, UpstreamRunsWhileDownstreamBackpressure) {
     fiber->await();
 }
 
+TEST(ObservableQueue, DownstreamStopBigQueue) {
+    int upstream_counter = 0;
 
+    auto sched = std::make_shared<BenchScheduler>();
+    auto downstream_queue = cask::Queue<int,cask::None>::empty(sched, 1);
+    auto fiber = Observable<int, cask::None>::fromVector({
+            0, 1, 2, 3, 4, 5
+        })
+        ->template map<int>([&upstream_counter](auto value) {
+            upstream_counter++;
+            return value;
+        })
+        ->queue(10)
+        ->take(2)
+        .run(sched);
+
+    sched->run_ready_tasks();
+    auto result = fiber->await();
+
+    EXPECT_EQ(upstream_counter, 6);
+    ASSERT_EQ(result.size(), 2);
+    EXPECT_EQ(result[0], 0);
+    EXPECT_EQ(result[1], 1);
+}
+
+TEST(ObservableQueue, DownstreamStopSmallQueue) {
+    int upstream_counter = 0;
+
+    auto sched = std::make_shared<BenchScheduler>();
+    auto downstream_queue = cask::Queue<int,cask::None>::empty(sched, 1);
+    auto fiber = Observable<int, cask::None>::fromVector({
+            0, 1, 2, 3, 4, 5
+        })
+        ->template map<int>([&upstream_counter](auto value) {
+            upstream_counter++;
+            return value;
+        })
+        ->queue(1)
+        ->take(2)
+        .run(sched);
+
+    sched->run_ready_tasks();
+    auto result = fiber->await();
+
+    EXPECT_EQ(upstream_counter, 2);
+    ASSERT_EQ(result.size(), 2);
+    EXPECT_EQ(result[0], 0);
+    EXPECT_EQ(result[1], 1);
+}